add A-packaging and refactoring of Task Variants
Changed the packaging conventions for the project to fit better A-Code quality.
