+++ src/main/java/Model/TaskVariants/Deadline.java
@@ -1,4 +1,4 @@
-package Model;
+package Model.TaskVariants;
 
 public class Deadline extends Task {
     protected String endDate;
