+++ src/main/java/Model/TaskVariants/Todo.java
@@ -1,4 +1,4 @@
-package Model;
+package Model.TaskVariants;
 
 public class Todo extends Task {
     public Todo(String taskDescription) {
